Modularising text values
