+++ src/app/config/configData.service.spec.ts
@@ -11,7 +11,7 @@ import {
 
 import {ResponseOptions} from '@angular/http';
 import {MockBackend, MockConnection} from '@angular/http/testing';
-import { InputComponent } from './inputComponent';
+import {InputComponent} from '../components/inputComponent';
 import { ConfigDataService } from './configData.service';
 
 describe('Job Service', () => {
